Update Apartment model to change parking type from byte to int; modify application properties for file profile; add ApartmentRepositoryTest for persistence testing

+++ src/main/java/com/cifo/apartmentpredictoralpy/model/Apartment.java
@@ -170,7 +170,7 @@ public class Apartment {
         return parking;
     }
 
-    public void setParking(byte parking) {
+    public void setParking(int parking) {
         this.parking = parking;
     }
